refactor: Runable state after rebasing
